refactor: introduce support.domain.persistence
