+++ src/main/java/com/meoguri/linkocean/domain/bookmark/persistence/CustomBookmarkRepositoryImpl.java
@@ -4,8 +4,8 @@ import static com.meoguri.linkocean.domain.bookmark.entity.QBookmark.*;
 import static com.meoguri.linkocean.domain.bookmark.entity.vo.BookmarkStatus.*;
 import static com.meoguri.linkocean.domain.profile.entity.QFollow.*;
 import static com.meoguri.linkocean.domain.tag.entity.QTag.*;
-import static com.meoguri.linkocean.support.querydsl.CustomPath.*;
-import static com.meoguri.linkocean.support.querydsl.JoinInfoBuilder.Initializer.*;
+import static com.meoguri.linkocean.support.domain.persistence.querydsl.CustomPath.*;
+import static com.meoguri.linkocean.support.domain.persistence.querydsl.JoinInfoBuilder.Initializer.*;
 import static com.querydsl.sql.SQLExpressions.*;
 import static org.apache.commons.lang3.BooleanUtils.*;
 
@@ -25,8 +25,8 @@ import com.meoguri.linkocean.domain.bookmark.persistence.dto.BookmarkFindCond;
 import com.meoguri.linkocean.domain.bookmark.persistence.dto.FindUsedTagIdWithCountResult;
 import com.meoguri.linkocean.domain.bookmark.persistence.dto.QFindUsedTagIdWithCountResult;
 import com.meoguri.linkocean.domain.profile.entity.Profile;
-import com.meoguri.linkocean.support.querydsl.CustomPath;
-import com.meoguri.linkocean.support.querydsl.Querydsl4RepositorySupport;
+import com.meoguri.linkocean.support.domain.persistence.querydsl.CustomPath;
+import com.meoguri.linkocean.support.domain.persistence.querydsl.Querydsl4RepositorySupport;
 import com.querydsl.core.BooleanBuilder;
 import com.querydsl.core.types.Order;
 import com.querydsl.core.types.OrderSpecifier;
