+++ src/main/java/com/meoguri/linkocean/support/domain/persistence/querydsl/Join.java
@@ -1,4 +1,4 @@
-package com.meoguri.linkocean.support.querydsl;
+package com.meoguri.linkocean.support.domain.persistence.querydsl;
 
 import java.util.List;
